+++ panel/services/__init__.py
@@ -1,0 +1,1 @@
+"""Long-running / domain services layered on top of ops and data."""